Added header guards to copy and move echoErroneousLine.hpp
