+++ src/header-utilities/echoErroneousLine.hpp
